// src/Web/CombatAnalysis.EnhancedWebApp/combatanalysis.enhancedwebapp.client/src/features/gameLogs/types/CombatPlayerStatsModel.ts
export type CombatPlayerStatsModel = {
    id: number;
    strength: number;
    agility: number;
    intelligence: number;
    stamina: number;
    spirit: number;
    dodge: number;
    parry: number;
    crit: number;
    haste: number;
    hit: number;
    expertise: number;
    armor: number;
    talents: string;
    combatId: number;
}